Move components up one level
