+++ src/dialog/styles/dialog.css.d.ts
