+++ src/slidePanel/createSlidePanelElement.ts
