+++ src/slidePanel/styles/slidePanel.css
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.css';
+@import '../../styles/variables.css';
 
 .underlay {
 	position: fixed;
