+++ src/slidePanel/styles/slidePanel.css.d.ts
