+++ tests/unit/all.ts
@@ -1,3 +1,3 @@
 import './main';
-import './components/createDialog';
-import './components/createSlidePanel';
+import './createDialog';
+import './createSlidePanel';
